Ensure all private modules are exported correctly (#2050)
