+++ packages/workbox-core/_private.mjs
@@ -7,26 +7,30 @@
 */
 
 // We either expose defaults or we expose every named export.
-import {DBWrapper} from './_private/DBWrapper.mjs';
-import {deleteDatabase} from './_private/deleteDatabase.mjs';
-import {WorkboxError} from './_private/WorkboxError.mjs';
 import {assert} from './_private/assert.mjs';
 import {cacheNames} from './_private/cacheNames.mjs';
 import {cacheWrapper} from './_private/cacheWrapper.mjs';
+import {DBWrapper} from './_private/DBWrapper.mjs';
+import {Deferred} from './_private/Deferred.mjs';
+import {deleteDatabase} from './_private/deleteDatabase.mjs';
+import {executeQuotaErrorCallbacks} from './_private/executeQuotaErrorCallbacks.mjs';
 import {fetchWrapper} from './_private/fetchWrapper.mjs';
 import {getFriendlyURL} from './_private/getFriendlyURL.mjs';
 import {logger} from './_private/logger.mjs';
+import {WorkboxError} from './_private/WorkboxError.mjs';
 
 import './_version.mjs';
 
 export {
-  DBWrapper,
-  deleteDatabase,
-  WorkboxError,
   assert,
   cacheNames,
   cacheWrapper,
+  DBWrapper,
+  Deferred,
+  deleteDatabase,
+  executeQuotaErrorCallbacks,
   fetchWrapper,
   getFriendlyURL,
   logger,
+  WorkboxError,
 };
