+++ packages/workbox-core/_private/cacheWrapper.mjs
@@ -6,13 +6,13 @@
   https://opensource.org/licenses/MIT.
 */
 
-import {pluginEvents} from '../models/pluginEvents.mjs';
-import {pluginUtils} from '../utils/pluginUtils.mjs';
 import {WorkboxError} from './WorkboxError.mjs';
 import {assert} from './assert.mjs';
-import {executeQuotaErrorCallbacks} from './quota.mjs';
 import {getFriendlyURL} from './getFriendlyURL.mjs';
 import {logger} from './logger.mjs';
+import {executeQuotaErrorCallbacks} from './executeQuotaErrorCallbacks.mjs';
+import {pluginEvents} from '../models/pluginEvents.mjs';
+import {pluginUtils} from '../utils/pluginUtils.mjs';
 import '../_version.mjs';
